+++ docs/CHANGELOG.md
@@ -1,9 +1,17 @@
 Changelog
 =========
 
-1.11.0-beta.8 (Unreleased)
+1.11.0-beta.9 (Unreleased)
+-----------------------------
+- Fix #444: Fix depreciation warnings on PHP 8.4
+
+1.11.0-beta.8 (July 22, 2025)
 -----------------------------
 - Enh #432: Add Yii formatter for Twig filters
+- Enh #437: Add context menu to Structure View
+- Enh #440: Add template section "CSS & JS"
+- Enh #441: Allow to render container edit wrapper attributes separately
+- Enh #442: Improve template structure view
 
 1.11.0-beta.7 (July 4, 2025)
 -----------------------------
@@ -28,6 +36,8 @@ Changelog
 
 1.11.0-beta.3 (June 3, 2025)
 ----------------------------
+Please note that the IDs of the custom pages you have created will be changed. Links may need to be updated.
+
 - Enh #354: Template - Add RSS Element Type
 - Enh #357: Refactor all snippets and container pages to single page class
 - Enh #358: Refactor page content types